+++ packages/fuels-test-helpers/src/lib.rs
@@ -3,6 +3,8 @@ extern crate core;
 
 use std::net::SocketAddr;
 
+#[cfg(feature = "fuels-accounts")]
+pub use accounts::*;
 #[cfg(feature = "fuel-core-lib")]
 pub use fuel_core::service::Config;
 #[cfg(feature = "fuel-core-lib")]
@@ -24,8 +26,6 @@ pub use node::*;
 #[cfg(not(feature = "fuel-core-lib"))]
 use portpicker::is_free;
 use rand::Fill;
-#[cfg(feature = "fuels-accounts")]
-pub use signers::*;
 #[cfg(feature = "fuel-core-lib")]
 pub use utils::{into_coin_configs, into_message_configs};
 pub use wallets_config::*;
@@ -34,7 +34,7 @@ pub use wallets_config::*;
 pub mod node;
 
 #[cfg(feature = "fuels-accounts")]
-mod signers;
+mod accounts;
 mod utils;
 mod wallets_config;
